March Update cycle.
This is the March update cycle. Still in MS3 Beta phase.

- Introduced the new Database middleware solution based on SQLITE (UNFINISHED, will not work yet as the endpoints are not all emulated)

- Introduced NautilusXP2024, a Home tools based on multiserver backend for much improved tool usability/performance.

- NetZ/Rendez-vous stability fixes, almost production ready (still the config to implement instead of hard-codded server list)

- Acorn API fixes for the Pik-nik minigame.

- General Horizon class fixup to get rid of duplicates/SQLITE Database comaptiblity.

: Amont of testing mid/high.

+++ BackendServices/AuxiliaryServices/HorizonUtils/LIBRARY.Database/Models/FileDTO.cs
@@ -10,10 +10,10 @@ namespace Horizon.LIBRARY.Database.Models
         public int CreationTimeStamp { get; set; }
         public int OwnerID { get; set; }
         public int GroupID { get; set; }
-        public ushort OwnerPermissionRWX { get; set; }
-        public ushort GroupPermissionRWX { get; set; }
-        public ushort GlobalPermissionRWX { get; set; }
-        public ushort ServerOperationID { get; set; }
+        public uint OwnerPermissionRWX { get; set; }
+        public uint GroupPermissionRWX { get; set; }
+        public uint GlobalPermissionRWX { get; set; }
+        public uint ServerOperationID { get; set; }
         public FileAttributesDTO? fileAttributesDTO { get; set; }
         public FileMetaDataDTO? fileMetaDataDTO { get; set; }
         public DateTime CreateDt { get; set; }
